Rename DeviceConnect to DeviceAction and drop DeviceInteraction
