+++ src/renderer/components/DeviceAction/actions/app.js
