+++ src/renderer/components/DeviceAction/actions/manager.js
