+++ src/renderer/components/DeviceAction/actions/shared.js
