+++ src/renderer/components/DeviceAction/animations.js
